Enhance list element.
List is provided as a table. If in one row, the number of cols is less
than the maximum number of cols, some col needs to increase the colspan.
Try to increase the colspan of a cell that has content (not just icons).

If a cell only contain icons, and no href is specified, but its left cell
contains content and has href set up, set the href of the cell to be the
same as the left cell.

Bump the version to 0.1.1